[#71630] Use bigint for ID column in friendly_id_slugs
Suggestion from @Copilot

+++ db/migrate/20260311000000_create_friendly_id_slugs.rb
@@ -32,7 +32,7 @@ class CreateFriendlyIdSlugs < ActiveRecord::Migration[8.1]
   def change
     create_table :friendly_id_slugs do |t|
       t.string   :slug,           null: false
-      t.integer  :sluggable_id,   null: false
+      t.bigint   :sluggable_id,   null: false
       t.string   :sluggable_type, limit: 50
       t.string   :scope
       t.datetime :created_at
